Add per-attribute schema handlers
